Allows selection of multihashes by name
Adds several new functions to the `Multihash` trait:
1. `Multihash::name(&self) -> &'static str`
2. `Multihash::new_with_name(name: &str, input: &[u8]) -> Result<Self, Error>`
3. `Multihash::wrap_with_name(name: &str, digest: &[u8]) -> Result<Self, Error>`

Together these functions create a (somewhat) nice interface for refering
to multihash hash algorithm variants by a more human-readable name.

Adds the `Error::UnsupportedName` variant to represent
invalid names in `Multihash::new_with_name` and
`Multihash::wrap_with_name`.

Adds the `#[mh(name = ...)]` attribute to the `Multihash` derive macro.
This allows the user to specify a custom name to associate with a given
hash algorithm variant. If this attribute is not specified, then the
stringified representation of the name of the variant corresponding to that hash
function will be used instead.

Adds nice error handeling & updated tests for the `#[mh(name)]`
attribute.

Resolves #32, although more work is probably needed for a full solution.