feat: rename `max_size` derive attribute to `alloc_size`
I originally choose the name `max_size` to give an indication what the user
should do with that value. Now that we have compile-time error messages I
think it makes sense to rename it to something that expresses more closely
what it really is about. It's about how many bytes are allocated.